requirements.txt
google-cloud-compute
google-api-core